Add files via upload
Our project makes use of 1 library – freeGLUT to achieve the desired graphics and animation. The following section gives overall idea of the objects, scenes and the method used for animating objects in the scene.

The project shows 6 scenes, Keyboard functions. The ‘X’ key is used to go to next scenes and the ‘Y’ key is used to exit the animation The introduction and the end scene is included in it.

The Scenario in the project is as follows:

We are showing an animation of how an object(i.e., car) finds the shortest path to reach its destination.